+++ screen/coordinates.go
@@ -5,9 +5,10 @@ package screen
 
 //------------------------------------------------------------------------------
 
-// Coord represents a pair of pixel coordinates.
+// Coord represents a pair of coordinates on the virtual screen, in (virtual)
+// pixels.
 type Coord struct {
-	X, Y int64
+	X, Y int16
 }
 
 //------------------------------------------------------------------------------
@@ -28,7 +29,7 @@ func (p Coord) Opposite() Coord {
 }
 
 // Times returns the product with a scalar.
-func (p Coord) Times(s int64) Coord {
+func (p Coord) Times(s int16) Coord {
 	return Coord{p.X * s, p.Y * s}
 }
 
@@ -39,7 +40,7 @@ func (p Coord) TimesCW(o Coord) Coord {
 
 // Slash returns the integer quotient of the division by a scalar (which must be
 // non-zero).
-func (p Coord) Slash(s int64) Coord {
+func (p Coord) Slash(s int16) Coord {
 	return Coord{p.X / s, p.Y / s}
 }
 
@@ -51,7 +52,7 @@ func (p Coord) SlashCW(o Coord) Coord {
 
 // Mod returns the remainder (modulus) of the division by a scalar (which must
 // be non-zero).
-func (p Coord) Mod(s int64) Coord {
+func (p Coord) Mod(s int16) Coord {
 	return Coord{p.X % s, p.Y % s}
 }
